requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.75.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
joblib==1.1.0
numpy==1.21.5
pickle-mixin==1.0.2
pydantic==1.9.0
python-aiml==0.9.3
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
sniffio==1.2.0
starlette==0.17.1
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
zipp==3.7.0